Renamed chart folder
