+++ packages/ketchup/src/components.d.ts
@@ -9,9 +9,6 @@ import '@stencil/core';
 
 
 import {
-  ChartConfig,
-} from './components/ketchup-chart/kup-chart-declarations';
-import {
   ComboItem,
 } from './components/ketchup-combo/ketchup-combo-declarations';
 import {
@@ -44,18 +41,12 @@ import {
 import {
   ButtonConfig,
 } from './components/kup-btn/kup-btn-declarations';
+import {
+  ChartConfig,
+} from './components/kup-chart/kup-chart-declarations';
 
 
 export namespace Components {
-
-  interface KupChart {
-    'config': ChartConfig;
-    'data': any;
-  }
-  interface KupChartAttributes extends StencilHTMLAttributes {
-    'config'?: ChartConfig;
-    'data'?: any;
-  }
 
   interface KetchupCombo {
     /**
@@ -588,6 +579,15 @@ export namespace Components {
     'transparent'?: boolean;
   }
 
+  interface KupChart {
+    'config': ChartConfig;
+    'data': any;
+  }
+  interface KupChartAttributes extends StencilHTMLAttributes {
+    'config'?: ChartConfig;
+    'data'?: any;
+  }
+
   interface MyComponent {
     /**
     * The first name
@@ -620,7 +620,6 @@ export namespace Components {
 
 declare global {
   interface StencilElementInterfaces {
-    'KupChart': Components.KupChart;
     'KetchupCombo': Components.KetchupCombo;
     'KupDash': Components.KupDash;
     'KupDataTable': Components.KupDataTable;
@@ -633,11 +632,11 @@ declare global {
     'KupTextInput': Components.KupTextInput;
     'KupBtn': Components.KupBtn;
     'KupButton': Components.KupButton;
+    'KupChart': Components.KupChart;
     'MyComponent': Components.MyComponent;
   }
 
   interface StencilIntrinsicElements {
-    'kup-chart': Components.KupChartAttributes;
     'ketchup-combo': Components.KetchupComboAttributes;
     'kup-dash': Components.KupDashAttributes;
     'kup-data-table': Components.KupDataTableAttributes;
@@ -650,15 +649,10 @@ declare global {
     'kup-text-input': Components.KupTextInputAttributes;
     'kup-btn': Components.KupBtnAttributes;
     'kup-button': Components.KupButtonAttributes;
+    'kup-chart': Components.KupChartAttributes;
     'my-component': Components.MyComponentAttributes;
   }
 
-
-  interface HTMLKupChartElement extends Components.KupChart, HTMLStencilElement {}
-  var HTMLKupChartElement: {
-    prototype: HTMLKupChartElement;
-    new (): HTMLKupChartElement;
-  };
 
   interface HTMLKetchupComboElement extends Components.KetchupCombo, HTMLStencilElement {}
   var HTMLKetchupComboElement: {
@@ -732,6 +726,12 @@ declare global {
     new (): HTMLKupButtonElement;
   };
 
+  interface HTMLKupChartElement extends Components.KupChart, HTMLStencilElement {}
+  var HTMLKupChartElement: {
+    prototype: HTMLKupChartElement;
+    new (): HTMLKupChartElement;
+  };
+
   interface HTMLMyComponentElement extends Components.MyComponent, HTMLStencilElement {}
   var HTMLMyComponentElement: {
     prototype: HTMLMyComponentElement;
@@ -739,7 +739,6 @@ declare global {
   };
 
   interface HTMLElementTagNameMap {
-    'kup-chart': HTMLKupChartElement
     'ketchup-combo': HTMLKetchupComboElement
     'kup-dash': HTMLKupDashElement
     'kup-data-table': HTMLKupDataTableElement
@@ -752,11 +751,11 @@ declare global {
     'kup-text-input': HTMLKupTextInputElement
     'kup-btn': HTMLKupBtnElement
     'kup-button': HTMLKupButtonElement
+    'kup-chart': HTMLKupChartElement
     'my-component': HTMLMyComponentElement
   }
 
   interface ElementTagNameMap {
-    'kup-chart': HTMLKupChartElement;
     'ketchup-combo': HTMLKetchupComboElement;
     'kup-dash': HTMLKupDashElement;
     'kup-data-table': HTMLKupDataTableElement;
@@ -769,6 +768,7 @@ declare global {
     'kup-text-input': HTMLKupTextInputElement;
     'kup-btn': HTMLKupBtnElement;
     'kup-button': HTMLKupButtonElement;
+    'kup-chart': HTMLKupChartElement;
     'my-component': HTMLMyComponentElement;
   }
