+++ packages/ketchup/src/components/kup-chart/kup-chart.spec.ts
